requirements.txt
click>=8
PyYAML>=6
platformdirs
rich
